Changing some smi dependencies to newer versions Modifying a few reference to a utility that was moved.
