+++ src/main/java/com/dell/isg/smi/service/server/inventory/controller/ServerLogController.java
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.dell.isg.smi.adapter.server.model.WsmanCredentials;
 import com.dell.isg.smi.commons.elm.exception.RuntimeCoreException;
-import com.dell.isg.smi.commons.elm.utilities.CustomRecursiveToStringStyle;
+import com.dell.isg.smi.commons.utilities.CustomRecursiveToStringStyle;
 import com.dell.isg.smi.commons.model.common.Credential;
 import com.dell.isg.smi.service.server.exception.BadRequestException;
 import com.dell.isg.smi.service.server.exception.EnumErrorCode;
